pyproject.toml
# Copyright 2024 Canonical Ltd.
# See LICENSE file for licensing details.

# Testing tools configuration
[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"
asyncio_mode = "auto"
markers = ["unstable"]

# Formatting tools configuration
[tool.black]
line-length = 99
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.poetry]
name = "spark-integration-hub-k8s"
version = "1.0"
description = "spark-integration-hub-k8s"
authors = []

[tool.poetry.dependencies]
python = ">3.8,<4.0"
ops = ">=2.10.1"
pyyaml = ">=6.0.1"
boto3 = ">=1.34.55"
jsonschema = "^4.19.1"
lightkube = "0.15.2"
rpds-py = "<0.19" # Newer versions of this rpds-py requires rustc 1.76 at the build time
                  # TODO: Unpin this once rustc 1.76 is available at build time


[tool.poetry.group.format]
optional = true

[tool.poetry.group.format.dependencies]
black = "^22.3.0"
ruff = ">=0.2.2"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
black = ">=22.3.0"
ruff = ">=0.2.2"
codespell = ">=2.2.2"

[tool.poetry.group.unit]
optional = true

[tool.poetry.group.unit.dependencies]
pytest = "<8.2.0"
pytest-asyncio = ">=0.21"
coverage = {extras = ["toml"], version = ">7.0"}
ops-scenario = "==5.8.1"

[tool.poetry.group.integration]
optional = true

[tool.poetry.group.integration.dependencies]
pytest = ">=7.2"
juju = "^3.5.2"
coverage = {extras = ["toml"], version = ">7.0"}
pytest-operator = ">0.20"
boto3 = ">=1.28.0"
tenacity = "^8.2.2"
pydantic ="^1.10.7"
spark8t = { git = "https://github.com/canonical/spark-k8s-toolkit-py.git", tag="v0.0.7"}

pytest-operator-cache = {git = "https://github.com/canonical/data-platform-workflows", tag = "v13.1.2", subdirectory = "python/pytest_plugins/pytest_operator_cache"}
# pytest-operator-groups = {git = "https://github.com/canonical/data-platform-workflows", tag = "v13.1.2", subdirectory = "python/pytest_plugins/pytest_operator_groups"}
# pytest-operator-secrets = {git = "https://github.com/canonical/data-platform-workflows", tag = "v13.1.2", subdirectory = "python/pytest_plugins/pytest_operator_secrets"}

[tool.ruff]
line-length = 99
select = ["E", "W", "F", "C", "N", "D", "I001"]
extend-ignore = [
    "D203",
    "D204",
    "D213",
    "D215",
    "D400",
    "D401",
    "D404",
    "D406",
    "D407",
    "D408",
    "D409",
    "D413",
]
ignore = ["E501", "D107"]
extend-exclude = ["__pycache__", "*.egg_info"]
per-file-ignores = {"tests/*" = ["D100","D101","D102","D103","D104", "E999"], "src/literals.py" = ["D101"]}
target-version="py310"
src = ["src", "tests"]

[tool.ruff.mccabe]
max-complexity = 10
